requirements.txt
discord.py==2.3.0
python-dotenv==1.0.0
PyNaCl==1.3.0
google-cloud-texttospeech==2.14.1
psycopg2-binary==2.9.6